Introduce SystemCog abstraction
